backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.25
botocore==1.34.25
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.26.0
aiohttp==3.9.1
beautifulsoup4==4.12.3
openai==1.10.0
langchain==0.1.4
langgraph==0.0.20
tiktoken==0.5.2
tenacity==8.2.3
structlog==24.1.0
python-dotenv==1.0.0
